Moved FASTOADModel class in problem.py
